Apply correctly change from master to 1.0.x
